Moved site codebase back up a directory.
The /webapp dir layout was screwing up our ability to keep the repo attached to github using a host that's particular about where public_html files are stored... We'll have to figure out another way of sharing the database. Don't really like the idea of keeping db files in the repo anyway.